requirements.txt
apache-beam
google-cloud-pubsub
avro
fastavro~=1.4.12
